==> 03_Scripts/1_Data_Wrangling.ipynb <==
# The Script wrangles and reformats data from the 'orders' and 'products' datasets. Then answers are provided to some ad-hoc client requests, and the wrangled data is exported

## The Script contains the following sections
### 1. Importing Libraries
### 2. Importing Data
### 3. Data Wrangling Steps
### 4. Reformatting Variables
### 5. Ad-hoc client requests
### 6. Exporting DFs


## 1. Importing Libraries

# Importing Libraries
import pandas as pd
import numpy as np
import os

## 2. Importing Data

# Creating shortcut to folder
path = r'C:\Users\seank\OneDrive\Dokumente\Career Foundry Data Analytics Course\Data Immersion\4 Python\03-2020_Instacart_Basket _Analysis'

# Creating list of variables to keep for orders df
orders_vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

# Importing datasets using path
     # Orders excluding eval_set
df_ords = pd.read_csv(os.path.join(path, '02_Data', 'Original_Data', 'orders.csv'), 
                            index_col = False,
                       usecols = orders_vars_list)
    # Product with all variables
df_prods = pd.read_csv(os.path.join(path, '02_Data', 'Original_Data', 'products.csv'), 
                            index_col = False)

    # Importing departments dataset
df_dep = pd.read_csv(os.path.join(path, '02_Data', 'Original_Data', 'departments.csv'),
                                   index_col = False)
df_dep.head()

## 3. Data Wrangling

# Transposing departments df
df_dep_t = df_dep.T

# Resetting Index
df_dep_t.reset_index()

# Changing the header to the first row
    # Extract new header
new_header = df_dep_t.iloc[0]
    # Deleting old first row
df_dep_t_new = df_dep_t[1:]
    # Assigning the header
df_dep_t_new.columns = new_header

# Turning dep df into data dictionary
data_dict = df_dep_t_new.to_dict('index')
data_dict

# Converting to df and saving
df_data_dict = pd.DataFrame.from_dict(data_dict, 
                                      orient="index")

# Saving dictionary
df_data_dict.to_pickle(os.path.join(path, "02_Data", "Prepared_Data", "dep_data_dict.pkl"))

# Renaming day of week variable
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# Changing order number to string
df_ords['order_id'] = df_ords['order_id'].astype('str')
df_ords['order_id'].dtype    # checking if it worked - it did

## 4. Reformatting Variables

Data Type Suitability

    # Investigating the variables
df_ords.info()    # user_id is an index that doesn't need to be integer
df_ords['user_id'] = df_ords['user_id'].astype('str')


Column Renaming Check

# Check current column names
df_ords.columns

## 5. Ad-hoc Client Requests

# Client wants to know what the busiest hour is for placing orders
busiest_hour = df_ords['order_hour_of_day'].value_counts().idxmax()  # Hour with max orders
order_count = df_ords['order_hour_of_day'].value_counts().max()  # Count of orders at busiest hour
print(f"The busiest hour for placing orders is {busiest_hour}:00 with {order_count} orders.")

# Checking overall distribution
df_ords['order_hour_of_day'].value_counts(dropna = False)

Sales Team wants a subset with breakfast item sales

# Locating Breakfast in the departments dictionary
data_dict    # it's 14

# Creating subset of all items in breakfast department
df_breakfast = df_prods[df_prods['department_id'] == 14]
df_breakfast.head()    # viewing first 5 rows

They want another subset of data with all dinner party items (alcohol, deli, beverages, and meat/seafood), and want to know how many rows in the dataframe

# Locating these departments in the data dictionary
data_dict    # alcohol = 5, deli = 20, beverages = 7, meat seafod = 12


df_prods['department_id'].dtype    # ensuring department_id is integer

 # subsetting
df_dinner_party = df_prods.loc[df_prods['department_id'].isin([5,20,7,12])]

# Counting rows of dinner party subset
num_of_rows = df_dinner_party.shape[0]
print(f"The number of rows is {num_of_rows}")

Someone from the data engineers team thinks they’ve spotted something strange about the customer with a "user_id" of “1.” Extracting all the information you can about this user.

# Subsetting for this user
df_user1 = df_ords.loc[df_ords['user_id'] == '1']
df_user1

# Basic stats on this user
df_user1.describe()  

user1modes = df_user1.mode().iloc[0] # also calculating most often occuring values
print(user1modes)

print('they have 11 total orders, \
most frequently ordered on Wednesday, \
Average time is 10am, \
with an average of 19 days between orders')

## 6. Exporting DFs

# exporting orders and departments dfs to csv files
df_ords.to_csv(os.path.join(path, '02_Data', 'Prepared_Data', 'orders_wrangled.csv'),
              index=False)

df_dep_t_new.to_csv(os.path.join(path, '02_Data', 'Prepared_Data', 'departments_wrangled.csv'),
                   index=False)
