==> 03_Scripts/2_Consistency Checks_dups_missing_dtype.ipynb <==
 # # Script cleaning and checking consistency of Orders and Products dataframes

### 1. Importing Libraries
### 2. Importing Data
### 3. Data Consistency Checks
		- Mixed Data Types
		- Missing Values
		- Trends and Outliers
		- Duplicates
### 4. Exporting Data

## 1. Importing Libraries

# Importing libraries
import pandas as pd
import numpy as np
import os

## 2. Importing Data

# Creating shortcut to folder
path = r'C:\Users\seank\OneDrive\Dokumente\Career Foundry Data Analytics Course\Data Immersion\4 Python\03-2020_Instacart_Basket _Analysis'

# Importing datasets using path
     # Orders dataset already prepared
df_ords = pd.read_csv(os.path.join(path, '02_Data', 'Prepared_Data', 'orders_wrangled.csv'), 
                            index_col = False)
     # Product 
df_prods = pd.read_csv(os.path.join(path,'02_Data', 'Original_Data', 'products.csv'),
                       index_col = False)

df_ords.head()

## 3. Data Consistency Checks

# Mixed Data Types

# Checking for mixed data types
for column in df_ords.columns:
    if df_ords[column].apply(type).nunique() > 1:
        print(f"Column '{column}' has mixed data types.")

no mixed data columns found

# Searching Out Missing Values

    # Products
# Using isnull (which results in boolean), and then adding up these results
df_prods.isnull().sum()

# creating subset with just the null values
df_prods_nan = df_prods[df_prods['product_name'].isnull() == True]
df_prods_nan

# Creating subset of data without the missings
df_prods.shape    # first check shape of data


# create subset with missings excluded
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]
df_prods_clean.shape     # check shape of this new df to make sure only 16 rows fewer

    # Now the Orders df
# checking for missings
df_ords.isnull().sum()

    Days_since_prior_order has 206209 missings
    This isn't an issue in cases where it's somebody's first order

# checking if these missings are all instances of people's first order
    # subset
df_ords_nan = df_ords[df_ords['days_since_prior_order'].isnull() == True]
df_ords_nan.describe()

Min and max for order number column are both 1, so this is not a true missing but valuable information since it represents people first orders. So we definitely shouldn't delete them.
Imputing values wouldn't make much sense either.
I could create a new binary variable to mark these orders as first order to be used in future analyses, but we can just use the order_number == 1 for this instead so it's unnecessary.
I will do nothing with these rows, and just stay aware of it when it comes to analysis stage

## Looking for trends and outliers

# Looking for other trends or outliers in Orders df
df_ords.describe()

The days of the week are coded as 0 - 6, rather than 1-7. Similarly time is 0-23 rather than 1-24.
We can also see that order_id is the unique identifier and that order number only goes until 100

# Now checking for trends in products df
df_prods_clean.describe()

# Max value of 99999 seems strange, investigating items with high prices
pricecheck = df_prods_clean[df_prods_clean['prices'] > 100]
pricecheck.head()

# It's just these two items that have been clearly miscoded. 
# Since I can't contact company, I will replace prices with prices of similar items

# Looking for cottage cheese and milk
# Filter rows where product_name contains 'cottage cheese' or 'milk'
milk_cheese_df = df_prods_clean[df_prods_clean['product_name'].str.contains('cottage cheese|milk', 
                                                                case=False, 
                                                                regex=True,
                                                               na=False)]

# Display the filtered df
milk_cheese_df

# it's clear that a lot of items contain Milk that aren't actually milk, like chocolate items.
# This rules out taking average or median of all 'milk' items
# Doing more specific search with milk as first word and also only 2% milk

df_2milk = milk_cheese_df[milk_cheese_df['product_name'].str.contains('^milk',
                                                                         case=False,
                                                                         regex=True,
                                                                         na=False) &
                            milk_cheese_df['product_name'].str.contains('2%',
                                                                         case=True,
                                                                         regex=True,
                                                                         na=False)
]

df_2milk.head(20)

# Taking average of these two items since they're  similar to our miscoded item
milk2avg = np.mean([2.7, 3.5])

# Reassign the price value for the miscoded milk (product_id 33664)
df_prods_clean.loc[df_prods_clean['product_id'] == 33664, 'prices'] = milk2avg

# Check it worked
print(df_prods_clean[df_prods_clean['product_id'] == 33664])

# Now for cottage cheese
    # subsetting all cottage cheeses
df_cottagecheese = df_prods_clean[df_prods_clean['product_name'].str.contains('cottage cheese',
                                                                         case=False,
                                                                         regex=True,
                                                                         na=False)]
# Now excluding the actual item with the issue
df_cottagecheese = df_cottagecheese[df_cottagecheese['product_id'] != 21553]

df_cottagecheese.describe()

df_cottagecheese.info()

There are 70 different items here so just taking the average of them to assign as value

# First calculate the average price in df_cottagecheese
average_price_cottage = df_cottagecheese['prices'].mean()
print(average_price_cottage)


# Assign the average price to the miscoded product (product_id is 21553)
df_prods_clean.loc[df_prods_clean['product_id'] == 21553, 'prices'] = average_price_cottage
print(df_prods_clean[df_prods_clean['product_id']==21553])

# Now checking how our changes have affected the main df
df_prods_clean.describe()

# # Finding and removing duplicates

# New df with full duplicates
    # products
df_prods_dups = df_prods_clean[df_prods_clean.duplicated()]
df_prods_dups

# Removing Duplicate
    # First checking size of df 
df_prods_clean.shape

# New Df without duplicates
df_prods_checked = df_prods_clean.drop_duplicates()
df_prods_checked.shape    # correct, only 5 fewer rows which is what we saw before

    # Now for Orders DF
# New df with full duplicates
df_ords_dups = df_ords[df_ords.duplicated()]
df_ords_dups

No duplicates so no action needed

# Renaming df to make clear it's been checked for consistency
df_ords_checked = df_ords

df_ords_checked.shape

## 4. Exporting to CSV

# orders
df_ords_checked.to_csv(os.path.join(path, '02_Data', 'Prepared_Data', 'orders_checked.csv'),
                       index = False)

# products
df_prods_checked.to_csv(os.path.join(path, '02_Data', 'Prepared_Data', 'products_checked.csv'),
                        index = False)
