==> 03_Scripts/3_Merge_orders_orders_prior.ipynb <==
## This script merges 'orders' and 'orders_prior' datasets and exports as pickle


### This script contains these sections

#### 1. Importing Libraries
#### 2. Importing Data
#### 3. Checking Dimensions of orders_prior data
#### 4. Merging Orders and Orders_Prior Datasets

## 1. Importing Libraries

# Importing libraries
import pandas as pd
import numpy as np
import os

## 2. Importing Data

# Creating shortcut to folder
path = r'C:\Users\seank\OneDrive\Dokumente\Career Foundry Data Analytics Course\Data Immersion\4 Python\03-2020_Instacart_Basket _Analysis'

# Importing datasets using path
     # Orders dataset already prepared
df_ords = pd.read_csv(os.path.join(path, '02_Data', 'Prepared_Data', 'orders_checked.csv'), 
                            index_col = False)

    # Raw Orders_Prior df
df_ords_prior = pd.read_csv(os.path.join(path, '02_Data', 'Original_Data', 'orders_products_prior.csv'),
                            index_col = False)

### 3. Checking Dimensions of Dataframes

# Investigating new df
df_ords_prior.info()

df_ords_prior.head()

df_ords_prior.describe()

# Checking shape of both DFs
df_ords_prior.shape

# Consistency Checks of orders Prior DF

# Missing Values
df_ords_prior.isnull().sum()

No missings

# Duplicates
df_ords_prior_dups = df_ords_prior[df_ords_prior.duplicated()]
df_ords_prior_dups

No Duplicates

df_ords.shape

# 4. Merging Data

# Merging two together on order id columns
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large.head()

df_merged_large['_merge'].value_counts()

# 4. Exporting as Pickle

# Export data to pkl
df_merged_large.to_pickle(os.path.join(path, '02_Data','Prepared_Data', 'orders_products_combined.pkl'))
