==> 03_Scripts/4_Merge_products_orders_prior.ipynb <==
# This script merges the orders_prior and product datasets

### This script contains these sections

#### 1. Importing Libraries
#### 2. Importing Data
#### 3. Checking Dimensions of data
#### 4. Merging Products and Orders_Prior Datasets
#### 5. Exporting Merged Dataset

# 1. Importing Libraries

# Importing libraries
import pandas as pd
import numpy as np
import os

# 2. Importing Data

# Creating shortcut to folder
path = r'C:\Users\seank\OneDrive\Dokumente\Career Foundry Data Analytics Course\Data Immersion\4 Python\03-2020_Instacart_Basket _Analysis'

# Importing Products Data that was already wrangled and checked
df_prods = pd.read_csv(os.path.join(path, '02_Data', 'Prepared_Data', 'products_checked.csv'),
                                 index_col = False)

# Importing Orders_prior data that was already merged 
df_ords_prods_comb = pd.read_pickle(os.path.join(path, '02_Data', 'Prepared_Data', 'orders_products_combined.pkl'))

# 3. Checking Dimensions of Data

df_ords_prods_comb.head()

# Check the shape of the imported dataframe 
df_ords_prods_comb.shape

This is the same as when I exported it, good

# Now checking cols of products before merging
df_prods.head()

# Double-Checking the issue with price is resolved in imported df
df_prods.describe()

# 4. Merging products and orders data

# since old merged column already had a merge flag, dropping this before new merge
df_ords_prods_comb.drop(columns=['_merge'], 
                        inplace=True)

# Using inner join since it was specified that we only want full data
    # Can use product_id since it has same name in both DFs
df_ords_prods = df_ords_prods_comb.merge(df_prods,
                                         on = 'product_id',
                                         how='inner',
                                         indicator=True)

df_ords_prods.head()

df_ords_prods.shape

# Checking results
df_ords_prods['_merge'].value_counts()

# Testing merge using outer to see if/how it differs
df_test = df_ords_prods_comb.merge(df_prods,
                                         on = 'product_id',
                                         how='outer',
                                         indicator=True)

df_test['_merge'].value_counts()

We can see that there were 30200 Records from the orders df and 11 from product that were left out of merge

# 5. Exporting combined data as Pickle

# Exporting as pickle since it's a very large file and next steps will be performed with python
df_ords_prods.to_pickle(os.path.join(path, '02_Data', 'Prepared_Data', 'ords_prods_merge.pkl'))
