==> 03_Scripts/5_Deriving_Variables.ipynb <==
# This script derives columns to categorise different aspects of the data.

# The script contains the following sections:
## 1. Importing Libraries
## 2. Importing Data
## 3. Deriving Columns
### 3a. Product Price-Range
### 3b. Busiest Days
### 3c. Busiest Hours
## 4. Exporting Data

## 1. Importing Libraries

# Importing libraries
import pandas as pd
import numpy as np
import os

## 2. Importing Data

# Creating shortcut to folder
path = r'C:\Users\seank\OneDrive\Dokumente\Career Foundry Data Analytics Course\Data Immersion\4 Python\03-2020_Instacart_Basket _Analysis'

# Importing DF with products and orders combined
ords_prods_merge = pd.read_pickle(os.path.join(path, '02_Data', 'Prepared_Data', 'ords_prods_merge.pkl'))
ords_prods_merge.head()

ords_prods_merge.describe()

# 3a. Deriving Price-Range Variable

# High Range Products
ords_prods_merge.loc[ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'

# Mid Range
ords_prods_merge.loc[(ords_prods_merge['prices'] <= 15) & (ords_prods_merge['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

# Low Range
ords_prods_merge.loc[ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

ords_prods_merge['price_range_loc'].value_counts()

# Above counted the orders of these products, here is counting the number of products in each category
ords_prods_merge[['product_id','price_range_loc']].drop_duplicates()['price_range_loc'].value_counts()

# 3b Deriving variable for busiest days

# checking which days are busiest
ords_prods_merge['orders_day_of_week'].value_counts()

# Creating variable categorising days according to busy-ness, one day version
result = []

for value in ords_prods_merge["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

result

# Appending results to the df
ords_prods_merge['busiest_day']=result

# Second variable that has multiple days

result2day = []

for value in ords_prods_merge['orders_day_of_week']:
    if value in [0,1]:
        result2day.append('Busiest days')
    elif value in [3,4]:
        result2day.append('Least busy days')
    else:
        result2day.append('Regularly busy')

# appending to df
ords_prods_merge['busiest_days'] = result2day

# checking new col
ords_prods_merge['busiest_days'].value_counts().to_clipboard(sep='\t')

It makes sense that the least busy days have the least total order-items. Regularly busy has more values than busiest days, but since that includes 3 weekdays rather than only 2 for 'busiest', this isn't alarming

# Creating a small table to see if the values go with the right days
# Create a small summary table
summary_table = ords_prods_merge[['orders_day_of_week', 'busiest_days']].value_counts().reset_index()


# Sort by count in descending order
summary_table = summary_table.sort_values('count', ascending=False)

# Display the summary
print(summary_table)

this looks as it should

# 3c. Deriving busiest hours column

# Checking which hours are busiest
ords_prods_merge['order_hour_of_day'].value_counts()

# Empty list 
bhours = []

# Assiging hours based on busiest 
    # based on interpretation of ords_prods_merge['order_hour_of_day'].value_counts() 
for value in ords_prods_merge['order_hour_of_day']:
    if 9 <= value <= 16: 
        bhours.append('Most orders')
    elif value in [17,8,18,19,20,7,21]:
        bhours.append('Average orders')
    else:
        bhours.append('Fewest orders')

# appending to df
ords_prods_merge['Busiest_hours'] = bhours

# Checking counts
ords_prods_merge['Busiest_hours'].value_counts()

ords_prods_merge['Busiest_hours'].value_counts().to_clipboard(sep='\t')

# 4. Exporting Data

# Exporting as Pickle
ords_prods_merge.to_pickle(os.path.join(path, '02_Data', 'Prepared_Data', 'ords_prods_vars.pkl'))
