==> 03_Scripts/6_Grouping and Aggregating Variables.ipynb <==
# Script that splits and flags customers based on loyalty, spending and frequency

# The script contains the following sections:
## 1. Importing Libraries
## 2. Importing Data
## 3. Deriving Columns
### 3a. Flagging number of orders
### 3b. Flagging spending (average item price)
### 3c. Flagging Frequency (median days between orders)
## 4. Exporting Data

# 1. Importing Libraries

# Importing libraries
import pandas as pd
import numpy as np
import os

# 2. Importing Data

# Folder Shortcut
path = r'C:\Users\seank\OneDrive\Dokumente\Career Foundry Data Analytics Course\Data Immersion\4 Python\03-2020_Instacart_Basket _Analysis'

# Importing Merged Data
ords_prods = pd.read_pickle(os.path.join(path, '02_Data', 'Prepared_Data', 'ords_prods_vars.pkl'))

# 3. Aggregating and Deriving Variables

# Average orders per user by department
ords_prods.groupby('department_id')['order_number'].mean().sort_values(ascending=False)

# Flagging customers based on number of orders

# column with customer's max orders
ords_prods['max_order'] = ords_prods.groupby(['user_id'])['order_number'].transform(np.max)

# Deriving 3-level 'loyalty_flag' var from max order score 
ords_prods.loc[ords_prods['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'
ords_prods.loc[(ords_prods['max_order'] <= 40) & (ords_prods['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'
ords_prods.loc[ords_prods['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# Checking output
ords_prods[['user_id', 'order_id', 'max_order', 'loyalty_flag']].head(100)

ords_prods[['user_id', 'loyalty_flag']].drop_duplicates()['loyalty_flag'].value_counts()

Checking if loyal customers are buying products that are different prices than regular/new customers

# Retrieving average, min, max, 1,3 quartile prices by customer group
    # Defining 25th Percentile
def q25(x):
    return x.quantile(0.25)

    # 75th Percentile
def q75(x):
    return x.quantile(0.75)

# Group by customer group and return some main statisitics about product prices
ords_prods.groupby('loyalty_flag').agg({'prices': ['mean', 'min', 'max', q25, q75]})

The products purchased by all customer groups have very similar average prices, as well as min, max and 1st and 3rd quartiles. The average price of the loyal customer group is actually slightly lower than the other groups

## Flagging customers as high/low spenders based on avg product price (<10<)

# Calculating average product price per customer
ords_prods['avg_price'] = ords_prods.groupby('user_id')['prices'].transform(np.mean)
ords_prods.head(100)

# Assigning customers as high or low spenders 
ords_prods.loc[ords_prods['avg_price'] < 10, 'spending_flag'] = 'Low spender'
ords_prods.loc[ords_prods['avg_price'] >= 10, 'spending_flag'] = 'High spender'

ords_prods[['user_id', 'order_id', 'avg_price', 'spending_flag']].head(100)    # checking output

# Counting values, making sure to first group by users so that's what's being counted
ords_prods[['user_id', 'spending_flag']].drop_duplicates()['spending_flag'].value_counts()

There are a lot more low spenders than high spenders

# Flagging customers' order frequency based on median days between orders

# Median days between orders for each customer
ords_prods['median_days_tween_orders'] = ords_prods.groupby('user_id')['days_since_prior_order'].transform('median')

# Check output
ords_prods[['user_id', 'order_id', 'median_days_tween_orders']]

# Checking if all users only have 1 value for new variable
ords_prods.groupby('user_id')['median_days_tween_orders'].nunique().value_counts()

# Finding this user with the 0
problem_user = ords_prods.groupby('user_id')['median_days_tween_orders'].nunique()
problem_user = problem_user[problem_user == 0]
print(problem_user.index)

ords_prods[ords_prods['user_id'].isin(problem_user.index)]
# They only have one order, hence NaN for days_since_prior_order

# Assigning flag based on median rules
ords_prods.loc[ords_prods['median_days_tween_orders'] > 20, 'frequency_flag'] = 'Non-frequent customer'
ords_prods.loc[(ords_prods['median_days_tween_orders'] > 10) & (ords_prods['median_days_tween_orders'] <= 20), 'frequency_flag'] = 'Regular customer'
ords_prods.loc[ords_prods['median_days_tween_orders'] <= 10, 'frequency_flag'] = 'Frequent customer'



# Checking frequencies
ords_prods['frequency_flag'].value_counts()

These are counts of the rows in the df (products within an order), whereas I want to count users

# Grouping by users
ords_prods[['user_id', 'frequency_flag']].drop_duplicates()['frequency_flag'].value_counts()

# Since so many frequent customers, checking if these are also customers with low amount of orders
    # Counting unique users by creating df with one user per row and their loyalty and frequency flag
user_freq_loyalty = ords_prods.groupby(['user_id', 'frequency_flag', 'loyalty_flag']).size().reset_index().drop(columns=0)

user_freq_loyalty[['frequency_flag', 'loyalty_flag']].value_counts().reset_index()

Even though frequent customers is the largest group, the loyal frequent customers is the smallest segment of this 

# 4. Exporting Data

# Exporting as Pickle
ords_prods.to_pickle(os.path.join(path, '02_Data', 'Prepared_Data', 'ords_prods_vars_flags.pkl'))
