==> 03_Scripts/7_Customer_wrangle_clean.ipynb <==
# Script that loads and performs checks on raw customer data then exports it

## The script contains the following sections
### 1. Importing Libraries
### 2. Importing Data
### 3. Wrangling Data
### 4. Consistency Checks
### 5. Exporting Checked Customer Data

# 1. Importing Libraries

# Importing libraries
import pandas as pd
import numpy as np
import os

# 2. Importing Data

# Folder Shortcut
path = r'C:\Users\seank\OneDrive\Dokumente\Career Foundry Data Analytics Course\Data Immersion\4 Python\03-2020_Instacart_Basket _Analysis'

# Importing Raw Customer Data
cust = pd.read_csv(os.path.join(path, '02_Data', 'Original_Data', 'customers.csv'),
                   index_col=False)

# 3. Data Wrangling

cust.head()    # previewing data

# Renaming variables
cust.rename(columns={'Surnam':'surname',
                      'First Name': 'first_name',
                      'Gender':'gender',
                      'Age':'age',
                      'STATE':'state'},
            inplace=True)

# Checking Variable Types
cust.info()

# Changing date to datetime format
cust['date_joined'] = pd.to_datetime(cust['date_joined'])
cust.dtypes

# Changing customer ID to string
cust['user_id'] = cust['user_id'].astype('str')

# examining values
cust.describe()

These max and mins look OK

# 4. Consistency Checks

Mixed Data Types

# Check for mixed data types
for column in cust.columns:
    if cust[column].apply(type).nunique() > 1:
        print(f"Column '{column}' has mixed data types.")

# Converting first name to string
cust['first_name'] = cust['first_name'].astype('str')

print(cust['first_name'].dtype)

Missing Values

cust.isnull().sum()

# but, we saw above  in ( cust.info() ) that first_name did have missing values. Checking if these were converted to nan
sorted_names = cust['first_name'].unique()
sorted_names = sorted(sorted_names)
print(sorted_names)

they were

# Recoding as Missing
cust['first_name'].replace(['nan'], pd.NA, 
                           inplace=True)

cust.isnull().sum()

Since it is only first name, we can do nothing here to these missings

# Duplicates

cust_dups = cust[cust.duplicated()]
cust_dups

No Duplicates found

## 5. Exporting Cleaned Checked customer Data


# Exporting as CSV
cust.to_csv(os.path.join(path, '02_Data', 'Prepared_Data', 'customers_checked.csv'),
            index = False)
