==> 03_Scripts/8_Merge_customer_ords_prods.ipynb <==
# Script that merges cleaned customer data with orders_products data and exports it as pickle

## Script contains following sections
### 1. Importing Libraries
### 2. Importing Data
### 3. Merging Data
### 4. Exporting Merged Data

# 1. Importing Libraries

# Importing libraries
import pandas as pd
import numpy as np
import os

# 2. Importing Data

# Folder Shortcut
path = r'C:\Users\seank\OneDrive\Dokumente\Career Foundry Data Analytics Course\Data Immersion\4 Python\03-2020_Instacart_Basket _Analysis'

# Importing Merged Orders Products Data
ords_prods = pd.read_pickle(os.path.join(path, '02_Data', 'Prepared_Data', 'ords_prods_vars_flags.pkl'))

# Importing Customer Data
cust = pd.read_csv(os.path.join(path, '02_Data', 'Prepared_Data', 'customers_checked.csv'),
                   index_col=False)

cust.head()

# Keep only necessary columns
cust = cust[['user_id', 'gender', 'state', 'age', 'n_dependants', 'fam_status','income']]

ords_prods.columns

# Dropping aisle_id, _merge, busiest_day, median_days_tween_orders
ords_prods = ords_prods[['order_id', 'user_id', 'order_number', 'orders_day_of_week',
                               'order_hour_of_day', 'days_since_prior_order', 'product_id', 
                               'add_to_cart_order', 'reordered', 'product_name', 
                               'department_id', 'prices', 'price_range_loc', 
                               'busiest_days', 'Busiest_hours', 'max_order', 
                               'loyalty_flag', 'avg_price', 'spending_flag', 'frequency_flag']]

# Changing variables types to reduce size
    # First making these both matching and string for merge
cust['user_id'] = cust['user_id'].astype('str')
ords_prods['user_id'] = ords_prods['user_id'].astype('str')

# Convert all int64 columns to int32
for col in cust.select_dtypes(include=['int64']).columns:
    cust[col] = cust[col].astype('int32')

for col in ords_prods.select_dtypes(include=['int64']).columns:
    ords_prods[col] = ords_prods[col].astype('int32')

# Reducing float64s to float32
for col in ords_prods.select_dtypes(include=['float64']).columns:
    ords_prods[col] = ords_prods[col].astype('float32')

ords_prods.dtypes

# Converting some columns to string in order products df
# List of columns to convert from int32 to string
columns_to_convert_op = ['order_id', 'order_day_of_week', 'product_id', 'department_id']  # Replace with your column names

# Convert the selected columns to string
for col in columns_to_convert_op:
    if col in ords_prods.columns:  # Ensure the column exists in the DataFrame
        ords_prods[col] = ords_prods[col].astype('string')

cust.dtypes

# 3. Merging Orders-Products with Customers

# Merging DFs on user_id
ords_prods_cust = cust.merge(ords_prods,
                             how='inner',
                             on='user_id',
                             indicator=True)

ords_prods_cust.head()

ords_prods_cust['_merge'].value_counts()

# 4. Exporting as Pickle

# Exporting as Pickle
ords_prods_cust.to_pickle(os.path.join(path, '02_Data', 'Prepared_Data', 'ords_prods_cust.pkl'))
